backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
cffi==1.16.0; python_version >= '3.8'
cryptography==42.0.7; python_version >= '3.7'
django==5.0.6; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-filter==24.2; python_version >= '3.8'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt[crypto]==5.3.1; python_version >= '3.8'
drf-spectacular[sidecar]==0.27.2; python_version >= '3.7'
drf-spectacular-sidecar==2024.5.1; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.22.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
psycopg[binary]==3.1.19; python_version >= '3.7'
psycopg-binary==3.1.19; python_version >= '3.7'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.35.1; python_version >= '3.8'
rpds-py==0.18.1; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
typing-extensions==4.11.0; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
